SwipeTableViewControllerを作成し、コードを見やすくした
